📦 NEW: Handle response stream helper function
