+++ packages/core/src/helpers/stream.ts
@@ -1,5 +1,6 @@
 import {ChatCompletionStream} from 'openai/lib/ChatCompletionStream';
 import {ChunkStream} from 'src/pipes';
+import {Stream} from 'openai/streaming';
 
 export interface Runner extends ChatCompletionStream<null> {}
 
@@ -32,3 +33,51 @@ export const getRunner = (readableStream: ReadableStream) => {
 export const getTextPart = (chunk: ChunkStream) => {
 	return chunk.choices[0]?.delta?.content || '';
 };
+
+/**
+ * Handles the response stream from a given `Response` object.
+ *
+ * @param {Object} params - The parameters for handling the response stream.
+ * @param {Response} params.response - The API response to handle.
+ * @param {boolean} params.rawResponse - Optional flag to include raw response headers.
+ *
+ * @returns {Object} An object containing the processed stream, thread ID, and optionally raw response headers.
+ * @returns {ReadableStream<any>} return.stream - The readable stream created from the response.
+ * @returns {string | null} return.threadId - The thread ID extracted from the response headers.
+ * @returns {Object} [return.rawResponse] - Optional raw response headers.
+ * @returns {Record<string, string>} return.rawResponse.headers - The headers from the raw response.
+ */
+export function handleResponseStream({
+	response,
+	rawResponse,
+}: {
+	response: Response;
+	rawResponse?: boolean;
+}): {
+	stream: any;
+	threadId: string | null;
+	rawResponse?: {
+		headers: Record<string, string>;
+	};
+} {
+	const controller = new AbortController();
+	const streamSSE = Stream.fromSSEResponse(response, controller);
+	const stream = streamSSE.toReadableStream();
+
+	const result: {
+		stream: ReadableStream<any>;
+		threadId: string | null;
+		rawResponse?: {
+			headers: Record<string, string>;
+		};
+	} = {
+		stream,
+		threadId: response.headers.get('lb-thread-id'),
+	};
+	if (rawResponse) {
+		result.rawResponse = {
+			headers: Object.fromEntries(response.headers.entries()),
+		};
+	}
+	return result;
+}
